@@ -4,7 +4,7 @@
     "type": "project",
     "license": "AGPL-3.0-or-later",
     "require": {
-        "php": ">=8.1, <9.0",
+        "php": ">=8.3, <9.0",
         "ext-curl": "*",
         "ext-iconv": "*",
         "ext-imagick": "*",
@@ -14,45 +14,46 @@
         "doctrine/annotations": "^2.0",
         "doctrine/doctrine-bundle": "^2.0",
         "doctrine/doctrine-migrations-bundle": "^3.2",
-        "doctrine/orm": "^2.4",
+        "doctrine/orm": "^3.5",
         "embed/embed": "^4.4",
         "firebase/php-jwt": "^6.3",
         "guzzlehttp/guzzle": "^7.0",
-        "nelmio/api-doc-bundle": "^4.0",
+        "nelmio/api-doc-bundle": "^5.4",
         "nelmio/cors-bundle": "^2.0",
-        "symfony/asset": "^6.2",
-        "symfony/console": "^6.2",
-        "symfony/dotenv": "^6.2",
-        "symfony/event-dispatcher": "^6.2",
-        "symfony/expression-language": "^6.2",
-        "symfony/flex": "^2.3",
-        "symfony/framework-bundle": "^6.2",
-        "symfony/intl": "^6.2",
-        "symfony/mailer": "^6.2",
-        "symfony/mime": "^6.2",
-        "symfony/monolog-bundle": "^3.0",
-        "symfony/password-hasher": "^6.2",
-        "symfony/polyfill-ctype": "^1.13",
-        "symfony/polyfill-uuid": "^1.18",
-        "symfony/property-info": "^6.2",
-        "symfony/rate-limiter": "^6.2",
-        "symfony/security-bundle": "^6.2",
-        "symfony/serializer": "^6.2",
-        "symfony/translation": "6.*",
-        "symfony/twig-bundle": "^6.2",
-        "symfony/validator": "~6.2",
-        "symfony/yaml": "~6.2",
-        "twig/extra-bundle": "^2.12|^3.0",
-        "twig/twig": "^2.12|^3.0"
+        "symfony/asset": "^7.3",
+        "symfony/console": "^7.3",
+        "symfony/dotenv": "^7.3",
+        "symfony/event-dispatcher": "^7.3",
+        "symfony/expression-language": "^7.3",
+        "symfony/flex": "^2.7",
+        "symfony/framework-bundle": "^7.3",
+        "symfony/intl": "^7.3",
+        "symfony/mailer": "^7.3",
+        "symfony/mime": "^7.3",
+        "symfony/monolog-bundle": "^3.10",
+        "symfony/password-hasher": "^7.3",
+        "symfony/polyfill-ctype": "^1.32",
+        "symfony/polyfill-uuid": "^1.32",
+        "symfony/property-info": "^7.3",
+        "symfony/rate-limiter": "^7.3",
+        "symfony/runtime": "^7.3",
+        "symfony/security-bundle": "^7.3",
+        "symfony/serializer": "^7.3",
+        "symfony/translation": "^7.3",
+        "symfony/twig-bundle": "^7.3",
+        "symfony/validator": "^7.3",
+        "symfony/yaml": "^7.3",
+        "twig/extra-bundle": "^3.21",
+        "twig/twig": "^3.21"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^3.16",
-        "phpstan/phpstan": "^1.0.1",
-        "phpunit/phpunit": "^10.3",
-        "symfony/browser-kit": "^6.2",
-        "symfony/css-selector": "^6.2",
-        "symfony/stopwatch": "^6.2",
-        "symfony/web-profiler-bundle": "^6.2"
+        "friendsofphp/php-cs-fixer": "^3.77",
+        "phpstan/phpstan": "^2.1.17",
+        "phpunit/phpunit": "^12.2",
+        "symfony/browser-kit": "^7.3",
+        "symfony/css-selector": "^7.3",
+        "symfony/stopwatch": "^7.3",
+        "symfony/web-profiler-bundle": "^7.3"
     },
     "config": {
         "preferred-install": {
@@ -61,7 +62,8 @@
         "sort-packages": true,
         "allow-plugins": {
             "composer/package-versions-deprecated": true,
-            "symfony/flex": true
+            "symfony/flex": true,
+            "symfony/runtime": true
         }
     },
     "autoload": {
@@ -108,7 +110,7 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "6.*",
+            "require": "7.*",
             "docker": false
         },
         "public-dir": "../public/api/",
