@@ -8,70 +8,49 @@ use OpenApi\Annotations as OA;
 use Symfony\Component\Serializer\Annotation\Groups;
 use Symfony\Component\Validator\Constraints as Assert;
 
-/**
- * @ORM\Table(name="`bookmark`")
- *
- * @ORM\Entity()
- */
+#[ORM\Entity]
+#[ORM\Table(name: '`bookmark`')]
 class Bookmark extends ApiEntity
 {
+    #[ORM\Id]
+    #[ORM\Column(type: 'guid')]
+    #[Assert\NotBlank]
+    #[Groups(['public'])]
     /**
-     * @Assert\NotBlank()
-     *
-     * @Groups("public")
-     *
      * @OA\Property(type="guid")
-     *
-     * @ORM\Column(type="guid")
-     *
-     * @ORM\Id
      */
-    private $id;
+    private string $id;
 
+    #[ORM\Column(type: 'integer')]
+    #[Assert\NotNull]
+    #[Assert\Type('integer')]
+    #[Groups(['public'])]
     /**
-     * @Assert\NotNull()
-     *
-     * @Assert\Type("integer")
-     *
-     * @Groups("public")
-     *
      * @OA\Property(type="integer")
-     *
-     * @ORM\Column(type="integer")
      */
-    private $createdAt;
+    private int $createdAt;
 
+    #[ORM\ManyToOne(targetEntity: User::class, inversedBy: 'bookmarks')]
+    #[Assert\NotNull]
+    #[Groups(['public'])]
     /**
-     * @Assert\NotNull()
-     *
-     * @Groups("public")
-     *
      * @OA\Property(type="App\Entity\User")
-     *
-     * @ORM\ManyToOne(targetEntity="App\Entity\User", inversedBy="bookmarks")
      */
-    private $user;
+    private User $user;
 
+    #[ORM\ManyToOne(targetEntity: Message::class, inversedBy: 'bookmarks')]
+    #[ORM\JoinColumn(nullable: false, onDelete: 'CASCADE')]
+    #[Groups(['read_bookmark'])]
     /**
-     * @ORM\ManyToOne(targetEntity="App\Entity\Message", inversedBy="bookmarks")
-     *
-     * @Groups({"read_bookmark"})
-     *
      * @OA\Property(type="App\Entity\Message")
      */
-    private $message;
+    private Message $message;
 
+    #[Groups(['public'])]
     /**
-     * @Groups("public")
-     *
      * @OA\Property(type="string")
      */
-    private $entityType;
-
-    public function getEntityType(): string
-    {
-        return strtolower((new \ReflectionClass($this))->getShortName());
-    }
+    private string $entityType;
 
     public function __construct()
     {
@@ -112,5 +91,10 @@ class Bookmark extends ApiEntity
     public function setMessage(Message $message): void
     {
         $this->message = $message;
+    }
+
+    public function getEntityType(): string
+    {
+        return strtolower((new \ReflectionClass($this))->getShortName());
     }
 }
