@@ -3,42 +3,36 @@
 namespace App\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Serializer\Annotation\Groups;
 use Symfony\Component\Validator\Constraints as Assert;
 
-/**
- * @ORM\Table(name="`system`")
- *
- * @ORM\Entity
- */
+#[ORM\Entity()]
+#[ORM\Table(name: "system")]
 class System
 {
     /**
-     * @ORM\Id
-     *
-     * @ORM\Column(type="string")
-     *
      * @Assert\Type("string")
      *
      * @Assert\NotBlank()
      */
+    #[ORM\Id]
+    #[ORM\Column(type: "guid")]
     private $key;
 
     /**
-     * @ORM\Column(type="text")
-     *
      * @Assert\Type("text")
      *
      * @Assert\NotNull()
      */
+    #[ORM\Column(type: "text")]
     private $value;
 
     /**
-     * @ORM\Column(type="integer")
-     *
      * @Assert\Type("integer")
      *
      * @Assert\NotNull()
      */
+    #[ORM\Column(type: "integer")]
     private $createdAt;
 
     public function getEntityType(): string
