@@ -2,7 +2,6 @@
 
 namespace App\Controller;
 
-use App\Normalizer\ObjectNormalizer;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
@@ -29,21 +28,20 @@ abstract class ApiController extends AbstractController
      */
     public function serialize(?object $object, array $groups = [])
     {
+        // Always add the 'public' group and ensure groups are unique.
+        $groups[] = 'public';
+        $groups = array_values(array_unique($groups));
+
         return $this->serializer->serialize(
             $object,
             'json',
             [
-                ObjectNormalizer::MAX_TREE_DEPTH_HANDLER => function ($object, $format = null, array $context = []) {
-                    return \method_exists($object, 'getId') ? ['id' => $object->getId()] : '';
-                },
                 AbstractNormalizer::CIRCULAR_REFERENCE_HANDLER => function ($object, $format = null, array $context = []) {
                     return \method_exists($object, 'getId') ? ['id' => $object->getId()] : '';
                 },
                 AbstractObjectNormalizer::SKIP_NULL_VALUES => true,
                 AbstractObjectNormalizer::CIRCULAR_REFERENCE_LIMIT => 1, // symfony default
                 AbstractObjectNormalizer::ALLOW_EXTRA_ATTRIBUTES => true, // symfony default
-                ObjectNormalizer::ENABLE_MAX_TREE_DEPTH => true,
-                ObjectNormalizer::TREE_DEPTH_LIMIT => 1,
                 'currentUser' => $this->getUser() ? $this->getUser()->getUserIdentifier() : null,
                 'groups' => $groups,
             ]
@@ -57,21 +55,20 @@ abstract class ApiController extends AbstractController
      */
     public function normalize($object, array $groups = [])
     {
+        // Always add the 'public' group and ensure groups are unique.
+        $groups[] = 'public';
+        $groups = array_values(array_unique($groups));
+
         return $this->serializer->normalize(
             $object,
             'json',
             [
-                ObjectNormalizer::MAX_TREE_DEPTH_HANDLER => function ($object, $format = null, array $context = []) {
-                    return \method_exists($object, 'getId') ? ['id' => $object->getId()] : '';
-                },
                 AbstractNormalizer::CIRCULAR_REFERENCE_HANDLER => function ($object, $format = null, array $context = []) {
                     return \method_exists($object, 'getId') ? ['id' => $object->getId()] : '';
                 },
                 AbstractObjectNormalizer::SKIP_NULL_VALUES => true,
                 AbstractObjectNormalizer::CIRCULAR_REFERENCE_LIMIT => 1, // symfony default
                 AbstractObjectNormalizer::ALLOW_EXTRA_ATTRIBUTES => true, // symfony default
-                ObjectNormalizer::ENABLE_MAX_TREE_DEPTH => true,
-                ObjectNormalizer::TREE_DEPTH_LIMIT => 1,
                 'currentUser' => $this->getUser() ? $this->getUser()->getUserIdentifier() : null,
                 'groups' => $groups,
             ]
