@@ -2,6 +2,8 @@
 
 namespace App\Service;
 
+use App\Entity\Notification as NotificationEntity;
+use App\Service\Notification as NotificationService;
 use App\Entity\User as UserEntity;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
@@ -10,13 +12,16 @@ class User
 {
     private $em;
     private $hasher;
+    private NotificationService $notificationService;
 
     public function __construct(
         EntityManagerInterface $em,
-        UserPasswordHasherInterface $hasher
+        UserPasswordHasherInterface $hasher,
+        NotificationService $notificationService
     ) {
         $this->em = $em;
         $this->hasher = $hasher;
+        $this->notificationService = $notificationService;
     }
 
     public function getByLogin($login)
@@ -54,5 +59,19 @@ class User
         $this->em->flush();
 
         return $user;
+    }
+
+    public function delete(UserEntity $user)
+    {
+        $notifications = $this->em->getRepository(NotificationEntity::class)->findBy(['owner' => $user,]);
+        foreach ($notifications as $notification) {
+            $this->notificationService->delete($notification);
+        }
+        $notifications = $this->em->getRepository(NotificationEntity::class)->findBy(['fromUser' => $user,]);
+        foreach ($notifications as $notification) {
+            $this->notificationService->delete($notification);
+        }
+        $this->em->remove($user);
+        $this->em->flush();
     }
 }
